update api find category update api delete category

+++ src/core/usecases/category/commands/delete-category/DeleteCategoryCommandHandler.ts
@@ -1,4 +1,5 @@
 import { ICategoryRepository } from '@gateways/repositories/category/ICategoryRepository';
+import { IProductRepository } from '@gateways/repositories/product/IProductRepository';
 import { MessageError } from '@shared/exceptions/message/MessageError';
 import { SystemError } from '@shared/exceptions/SystemError';
 import { CommandHandler } from '@shared/usecase/CommandHandler';
@@ -11,10 +12,18 @@ export class DeleteCategoryCommandHandler implements CommandHandler<string, Dele
     @Inject('category.repository')
     private readonly _categoryRepository: ICategoryRepository;
 
+    @Inject('product.repository')
+    private readonly _productRepository: IProductRepository;
+
     async handle(id: string): Promise<DeleteCategoryCommandOutput> {
         const category = await this._categoryRepository.getById(id);
         if (!category)
             throw new SystemError(MessageError.PARAM_NOT_EXISTS, 'category');
+
+        const ids = await this._categoryRepository.findTreeById(id);
+        const isExistProduct = await this._productRepository.checkExistByCategoryIds(ids);
+        if (isExistProduct)
+            throw new SystemError(MessageError.OTHER, 'Danh mục đã tồn tại sản phẩm không thể xóa');
 
         const hasSucceed = await this._categoryRepository.softDelete(id);
         const result = new DeleteClientCommandOutput();
